// vanila_redux_todo-list/src/enteties/episode/model/actionTypes/episodeActionTypes.ts
export enum EpisodeActionTypes {
  SET_IS_LOADING = "EPISODE/SET_IS_LOADING",
  SET_ERROR = "EPISODE/SET_ERROR",
  SET_EPISODES = "EPISODE/SET_EPISODE",
  SET_SINGLE_EPISODE = "EPISODE/SET_SINGLE_EPISODE",
  CLEAR_ALL_EPISODE_META = "EPISODE/CLEAR_ALL_EPISODE_META ",
  SET_EPISODE_SEARCH_QUERY = "EPISODE/SET_EPISODE_SEARCH_QUERY",
  SET_EPISODE_PAGE = "EPISODE/SET_EPISODE_PAGE",
  SET_EPISODE_PAGE_FORWARD = "EPISODE/SET_EPISODE_PAGE_FORWARD",
  SET_EPISODE_PAGE_BACK = "EPISODE/SET_EPISODE_PAGE_BACK",
}